Added: added removing programming interface
- [Add] Added removing programming interface
- [Change] changed mermaid graph from `TD` to `LR` in README.md
