+++ src/comments/getInlineExclude.ts
@@ -1,9 +1,9 @@
+import { getCommentNamespaces } from '#/comments/getCommentNamespaces';
 import { getJsDocComment } from '#/comments/getJsDocComment';
 import { getJsDocTag } from '#/comments/getJsDocTag';
 import type { IInlineExcludeInfo } from '#/comments/interfaces/IInlineExcludeInfo';
 import type { IStatementComments } from '#/comments/interfaces/IStatementComments';
 import { parse } from 'comment-parser';
-import { getCommentNamespaces } from './getCommentNamespaces';
 
 export function getInlineExclude(params: {
   comment: IStatementComments;
